0.1 Functional App. To-Do: -- Fix SearchCityByName -- Add icons -- Add images

+++ app/src/main/java/com/juanimozo/weatherapp/data/network/accuweather_api/current_conditions/MetricXXXXXXXXXXXXXXXXX.kt
@@ -3,5 +3,5 @@ package com.juanimozo.weatherapp.data.network.accuweather_api.current_conditions
 data class MetricXXXXXXXXXXXXXXXXX(
     val Unit: String,
     val UnitType: Int,
-    val Value: Int
+    val Value: Double
 )
