+++ app/src/main/java/com/juanimozo/weatherapp/data/network/accuweather_api/weekly_forecast/Maximum.kt
@@ -3,5 +3,5 @@ package com.juanimozo.weatherapp.data.network.accuweather_api.weekly_forecast
 data class Maximum(
     val Unit: String,
     val UnitType: Int,
-    val Value: Int
+    val Value: Double
 )
